backend/pyproject.toml
[project]
name = "marsa-backend"
version = "0.1.0"
description = "MARSA backend service"
requires-python = ">=3.12"
dependencies = [
	"fastmcp>=0.1.0",
	"langgraph>=0.2.0",
	"langchain-core>=0.3.0",
	"langchain-anthropic>=0.2.0",
	"chromadb>=0.5.0",
	"tavily-python>=0.4.0",
	"fastapi>=0.115.0",
	"uvicorn>=0.32.0",
	"python-dotenv>=1.0.0",
	"pytest>=8.0.0",
	"pytest-asyncio>=0.24.0",
	"httpx>=0.27.0",
	"pydantic>=2.9.0",
	"openai>=1.54.0",
	"tenacity>=9.0.0",
	"structlog>=24.4.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
	"integration: marks tests as integration tests that use real APIs (deselect with '-m \"not integration\"')"
]
